Move all stimuli into `stimuli` subdir (though import up)
