+++ stimupy/stimuli/benarys.py
