+++ stimupy/stimuli/checkerboards.py
