+++ stimupy/stimuli/cornsweets.py
