+++ stimupy/stimuli/cubes.py
