+++ stimupy/stimuli/gabors.py
