+++ stimupy/stimuli/mondrians.py
