+++ stimupy/stimuli/mueller_lyers.py
