+++ stimupy/stimuli/pinwheels.py
