+++ stimupy/stimuli/plaids.py
