+++ stimupy/stimuli/todorovics.py
