+++ stimupy/stimuli/waves.py
