+++ stimupy/stimuli/wedding_cakes.py
